primitives can move
a function has been written to calculate the inertia tensor of a primitive given a spatial transformation. Work has also begun on a visualiser for verification tests and further uses.
