+++ Documentation/dev_vault/.obsidian/workspace.json
@@ -16,7 +16,9 @@
                 "file": "Rocket Definition/Primitives/Conic Generalised.md",
                 "mode": "source",
                 "source": false
-              }
+              },
+              "icon": "lucide-file",
+              "title": "Conic Generalised"
             }
           },
           {
@@ -28,7 +30,9 @@
                 "file": "Rocket Definition/Primitives/Conic Full.md",
                 "mode": "source",
                 "source": false
-              }
+              },
+              "icon": "lucide-file",
+              "title": "Conic Full"
             }
           },
           {
@@ -40,21 +44,39 @@
                 "file": "Kinematics/Reference Frames.md",
                 "mode": "source",
                 "source": false
-              }
+              },
+              "icon": "lucide-file",
+              "title": "Reference Frames"
             }
           },
           {
             "id": "437b4d6ace5e7f9c",
             "type": "leaf",
             "state": {
+              "type": "markdown",
+              "state": {
+                "file": "Rocket Definition/Primitives/Primitives.md",
+                "mode": "preview",
+                "source": false
+              },
+              "icon": "lucide-file",
+              "title": "Primitives"
+            }
+          },
+          {
+            "id": "e878703e66129545",
+            "type": "leaf",
+            "state": {
               "type": "release-notes",
               "state": {
-                "currentVersion": "1.6.7"
-              }
+                "currentVersion": "1.7.4"
+              },
+              "icon": "lucide-file",
+              "title": "Release Notes 1.7.4"
             }
           }
         ],
-        "currentTab": 2
+        "currentTab": 4
       }
     ],
     "direction": "vertical"
@@ -74,7 +96,9 @@
               "type": "file-explorer",
               "state": {
                 "sortOrder": "alphabetical"
-              }
+              },
+              "icon": "lucide-folder-closed",
+              "title": "Files"
             }
           },
           {
@@ -89,7 +113,9 @@
                 "collapseAll": false,
                 "extraContext": false,
                 "sortOrder": "alphabetical"
-              }
+              },
+              "icon": "lucide-search",
+              "title": "Search"
             }
           },
           {
@@ -97,7 +123,9 @@
             "type": "leaf",
             "state": {
               "type": "bookmarks",
-              "state": {}
+              "state": {},
+              "icon": "lucide-bookmark",
+              "title": "Bookmarks"
             }
           }
         ]
@@ -120,7 +148,7 @@
             "state": {
               "type": "backlink",
               "state": {
-                "file": "Kinematics/Reference Frames.md",
+                "file": "Rocket Definition/Primitives/Conic Generalised.md",
                 "collapseAll": false,
                 "extraContext": false,
                 "sortOrder": "alphabetical",
@@ -128,7 +156,9 @@
                 "searchQuery": "",
                 "backlinkCollapsed": false,
                 "unlinkedCollapsed": true
-              }
+              },
+              "icon": "links-coming-in",
+              "title": "Backlinks for Conic Generalised"
             }
           },
           {
@@ -137,10 +167,12 @@
             "state": {
               "type": "outgoing-link",
               "state": {
-                "file": "Kinematics/Reference Frames.md",
+                "file": "Rocket Definition/Primitives/Conic Generalised.md",
                 "linksCollapsed": false,
                 "unlinkedCollapsed": true
-              }
+              },
+              "icon": "links-going-out",
+              "title": "Outgoing links from Conic Generalised"
             }
           },
           {
@@ -151,7 +183,9 @@
               "state": {
                 "sortOrder": "frequency",
                 "useHierarchy": true
-              }
+              },
+              "icon": "lucide-tags",
+              "title": "Tags"
             }
           },
           {
@@ -160,8 +194,10 @@
             "state": {
               "type": "outline",
               "state": {
-                "file": "Kinematics/Reference Frames.md"
-              }
+                "file": "Rocket Definition/Primitives/Conic Generalised.md"
+              },
+              "icon": "lucide-list",
+              "title": "Outline of Conic Generalised"
             }
           }
         ]
@@ -181,16 +217,16 @@
       "command-palette:Open command palette": false
     }
   },
-  "active": "c00fce086dce448c",
+  "active": "e878703e66129545",
   "lastOpenFiles": [
-    "Kinematics/Frame Transformation.md",
-    "Kinematics/Reference Frames.md",
-    "Rocket Definition/Primitives/Conic Generalised.md",
-    "Rocket Definition/Primitives/Conic Full.md",
     "Rocket Definition/Primitives/Custom Primitive.md",
+    "Rocket Definition/Primitives/Conic Generalised.md",
     "Rocket Definition/Primitives/Primitives.md",
-    "Kinematics/Chapter 4. 3D Rotations.md",
     "Rocket Definition/Primitives/Rectangular Prism.md",
+    "Rocket Definition/Primitives/Conic Full.md",
+    "Kinematics/Reference Frames.md",
+    "Kinematics/Frame Transformation.md",
+    "Kinematics/Chapter 4. 3D Rotations.md",
     "Kinematics",
     "Rocket Definition/Coordinate Systems.md",
     "Rocket Definition",
