--- visualiser.py
@@ -1,1 +1,0 @@
-# Show 3d objects as wireframes using openGL
